data-pipeline/dcn_dashboard/requirements.txt
streamlit==1.31.1
pandas==2.0.3
numpy==1.24.3
plotly==5.18.0
pyarrow==14.0.1
fastparquet==2024.2.0
python-swiftclient==4.4.0
python-keystoneclient==5.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9 